requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2
django-filter==22.1
djangorestframework==3.12.4
idna==3.4
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2022.7.1
requests==2.28.2
schedule==1.2.0
sqlparse==0.4.3
urllib3==1.26.15
